+++ package.json
@@ -68,6 +68,7 @@
     "@sentry/react": "^7.80.1",
     "@twilio/conversations": "^2.4.1",
     "@types/inquirer": "^9.0.7",
+    "aws-cdk": "2.177.0",
     "aws-cdk-lib": "^2.177.0",
     "aws-lambda": "^1.0.7",
     "browser-image-compression": "^2.0.2",
