--- scripts/deploy-test/deploy-config.json
@@ -1,6 +1,0 @@
-{
-  "project_id": "your_project_id",
-  "access_token": "your_access_token",
-  "provider_email": "your_provider_email",
-  "environment": "your_environment_name"
-}
