+++ scripts/deploy/.gitignore
@@ -1,0 +1,1 @@
+deploy-config.json
