+++ scripts/deploy/aws/.npmignore
