+++ scripts/deploy/gcp/.gitignore
@@ -1,0 +1,5 @@
+.terraform
+*.tfstate
+*.tfstate.*
+*.tfvars
+*.tfvars.json
