+++ scripts/deploy/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "deploy-test",
+  "name": "deploy",
   "version": "0.1.0",
   "bin": {
-    "deploy-test": "bin/deploy-test.js"
+    "deploy": "bin/deploy.js"
   },
   "engines": {
     "node": ">=20.0.0",
@@ -26,9 +26,11 @@
   "dependencies": {
     "@aws-sdk/client-cloudfront": "^3.645.0",
     "@aws-sdk/credential-providers": "^3.645.0",
+    "@oystehr/sdk": "3.0.7",
     "@types/node-fetch": "^2.6.11",
     "aws-cdk-lib": "2.177.0",
     "constructs": "^10.0.0",
+    "ts-node": "10.9.2",
     "source-map-support": "^0.5.21"
   }
 }
